+++ src/CLI/Drop.php
@@ -20,11 +20,8 @@ class Drop implements Command {
      * @return boolean
      */
     public function run( array $args = [], array $assoc_args = [] ) : bool {
-        if ( count( $args ) === 0 ) {
-            return $this->drop_index( 'posts' );
-        }
-        elseif ( count( $args ) === 1 ) {
-            return $this->drop_index( $args[0] );
+        if ( count( $args ) === 1 ) {
+            return $this->drop_index( $args[0], $assoc_args );
         }
         elseif ( count( $args ) > 1 ) {
             WP_CLI::error( 'RediPress: "drop" command does not accept more than two parameters.' );
@@ -32,29 +29,38 @@ class Drop implements Command {
         }
     }
 
-    public function drop_index( string $index ) {
+    /**
+     * Drop the index
+     *
+     * @param string $index The index to delete.
+     * @param array $assoc_args The associative args.
+     * @throws \Exception If index type is not supported.
+     * @return bool
+     */
+    public function drop_index( string $index, array $assoc_args ) {
         switch ( $index ) {
             case 'posts':
-                $return = apply_filters( 'redipress/drop_index', null );
+                $return = apply_filters( 'redipress/index/posts/drop', 'method not found', $assoc_args );
                 break;
             case 'users':
-                $return = apply_filters( 'redipress/drop_user_index', null );
+                $return = apply_filters( 'redipress/index/users/drop', 'method not found', $assoc_args );
                 break;
             default:
                 throw new \Exception( 'Index type ' . $index . ' is not supported.' );
                 break;
         }
 
-        switch ( $return ) {
-            case true:
-                WP_CLI::success( 'Index deleted.' );
-                return true;
-            case 'Unknown Index name':
-                WP_CLI::error( 'There was no index to delete or it was created under another name.' );
-                return false;
-            default:
-                WP_CLI::error( 'Unprecetended response: ' . $return );
-                return false;
+        if ( $return === true ) {
+            \WP_CLI::success( 'Index deleted.' );
+            return true;
+        }
+        elseif ( $return === 'Unknown Index name' ) {
+            \WP_CLI::error( 'There was no index to delete or it was created under another name.' );
+            return false;
+        }
+        else {
+            \WP_CLI::error( 'Unprecetended response: ' . $return );
+            return false;
         }
     }
 
@@ -64,7 +70,7 @@ class Drop implements Command {
      * @return integer
      */
     public static function get_min_parameters() : int {
-        return 0;
+        return 1;
     }
 
     /**